printer: move out of models to allow for other printer types
